Add GDrive authentication and cleanup query handling

+++ src/GDriveCloudStorage.js
@@ -1,10 +1,15 @@
 import GDrive from 'react-native-google-drive-api-wrapper';
 import { GoogleSignin } from '@react-native-community/google-signin';
 
-async function authenticate() {
+export async function authenticate(options = {}) {
   GoogleSignin.configure({
     scopes: ['https://www.googleapis.com/auth/drive.appdata'],
+    ...options,
   });
+  await GoogleSignin.hasPlayServices({
+    showPlayServicesUpdateDialog: true,
+  });
+  await GoogleSignin.signIn();
   const { accessToken } = await GoogleSignin.getTokens();
   GDrive.setAccessToken(accessToken);
   GDrive.init();
@@ -14,30 +19,35 @@ async function authenticate() {
 }
 
 export async function setItem(keyId, value) {
-  await initialGoogle();
   return GDrive.files.createFileMultipart(
     '',
     'text/plain',
     {
       parents: ['appDataFolder'],
       name: keyId,
-      appProperties: {
-        keyId: value,
-      },
+      appProperties: { value },
     },
     false,
   );
 }
 
+async function _getFileId(keyId) {
+  return GDrive.files.getId(keyId, ['appDataFolder'], 'text/plain', false);
+}
+
 export async function getItem(keyId) {
-  await initialGoogle();
-  const fileId = await GDrive.files.getId(keyId, ['appDataFolder'], 'text/plain', false);
+  const fileId = await _getFileId(keyId);
+  if (!fileId) {
+    return null;
+  }
   const meta = await GDrive.files.get(fileId);
-  return JSON.stringify(meta.appProperties);
+  return meta ? meta.appProperties.value : null;
 }
 
 export async function removeItem(keyId) {
-  await initialGoogle();
-  const fileId = await GDrive.files.getId(keyId, ['appDataFolder'], 'text/plain', false);
+  const fileId = await _getFileId(keyId);
+  if (!fileId) {
+    throw new Error('Item not found');
+  }
   return GDrive.files.delete(fileId);
 }
